Remove unused test, reorganize Input fn
