+++ lib/console/board.ex
@@ -15,16 +15,6 @@ defmodule TicTacToe.Console.Board do
     |> Enum.join("")
   end
 
-  def get_square_for_turn(player) do
-    current_player =
-      Atom.to_string(player)
-      |> String.capitalize()
-
-    IO.gets("Player #{current_player}. Choose a square: ")
-    |> String.trim("\n")
-    |> String.to_integer()
-  end
-
   def print(board) do
     board_string = format(board)
     IO.puts(board_string)
