+++ lib/tic_tac_toe.ex
@@ -1,11 +1,12 @@
 defmodule TicTacToe do
   alias TicTacToe.Core.Game, as: Game
   alias TicTacToe.Console.Board, as: Board
+  alias TicTacToe.Console.Input, as: Input
 
   def main(_) do
     game = Game.new()
     Board.print(game.board)
-    move = Board.get_square_for_turn()
+    move = Input.get_square_for_turn(game.next_player)
     updated_game = Game.take_turn(game, move)
     Board.print(updated_game.board)
     updated_game
