+++ test/console/board_test.exs
@@ -1,6 +1,5 @@
 defmodule TicTacToe.Console.BoardTest do
   use ExUnit.Case
-  import ExUnit.CaptureIO
 
   alias TicTacToe.Console.Board, as: ConsoleBoard
   alias TicTacToe.Core.Board, as: Board
@@ -37,47 +36,6 @@ defmodule TicTacToe.Console.BoardTest do
 
       board_string = ConsoleBoard.format(board)
       assert "O X 3 \nO X 6 \n7 8 X " == board_string
-    end
-  end
-
-  describe "get_square_for_turn" do
-    @describetag TicTacToeConsoleBoard: "get_square_for_turn"
-
-    test "presents player X with a prompt" do
-      player = :x
-
-      string_player =
-        Atom.to_string(player)
-        |> String.capitalize()
-
-      expected = "Player #{string_player}. Choose a square: "
-
-      assert expected ==
-               capture_io(:stdio, [input: "5", capture_prompt: true], fn ->
-                 ConsoleBoard.get_square_for_turn(player)
-               end)
-    end
-
-    test "presents player O with a prompt" do
-      player = :o
-
-      string_player =
-        Atom.to_string(player)
-        |> String.capitalize()
-
-      expected = "Player #{string_player}. Choose a square: "
-
-      assert expected ==
-               capture_io(:stdio, [input: "5", capture_prompt: true], fn ->
-                 ConsoleBoard.get_square_for_turn(player)
-               end)
-    end
-
-    test "returns the input value as an integer" do
-      capture_io(:stdio, [input: "5", capture_prompt: true], fn ->
-        value = ConsoleBoard.get_square_for_turn(:x)
-        assert 5 == value
-      end)
     end
   end
 end
