--- test/tic_tac_toe_test.exs
@@ -1,7 +1,0 @@
-defmodule TicTacToeTest do
-  use ExUnit.Case
-  import ExUnit.CaptureIO
-
-  alias TicTacToe.Core.Game, as: Game
-  @moduletag :TicTacToe
-end
